requirements.txt
numpy
pandas
scikit-learn
openai
google-generativeai
requests